requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
requests==2.31.0
APScheduler==3.10.4
